analysis_part/requirements.txt
pandas
numpy
scikit-learn
statsmodels
joblib
google-cloud-storage
xgboost